last push

+++ ex03/Point.cpp
@@ -6,7 +6,7 @@
 /*   By: ychahbi <ychahbi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/15 17:51:30 by ychahbi           #+#    #+#             */
-/*   Updated: 2023/12/17 11:18:43 by ychahbi          ###   ########.fr       */
+/*   Updated: 2023/12/27 14:55:26 by ychahbi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,13 +29,13 @@ Point::Point(const Fixed& x, const Fixed& y) : x(x), y(y) {}
 
 float   Point::get_x() const
 {
-    float Px = this->x.toInt();
+    float Px = this->x.toFloat();
     return (Px);
 }
 
 float   Point::get_y() const
 {
-    float Py = this->y.toInt();
+    float Py = this->y.toFloat();
     return (Py);
 }
